+++ kb/config.py
@@ -1,5 +1,5 @@
 # -*- encoding: utf-8 -*-
-# kb v0.1.3
+# kb v0.1.4
 # A knowledge base organizer
 # Copyright © 2020, gnc.
 # See /LICENSE for licensing information.
@@ -23,18 +23,16 @@ DEFAULT_CONFIG = {
     "PATH_KB_DB": str(Path(Path.home(), ".kb", "kb.db")),
     "PATH_KB_HIST": str(Path(Path.home(), ".kb", "recent.hist")),
     "PATH_KB_DATA": str(Path(Path.home(), ".kb", "data")),
-    # for future use
-    "PATH_KB_CONFIG": str(Path(Path.home(), ".kb", "kb.conf.py")),
-    "PATH_KB_MARKERS": str(Path(Path.home(), ".kb", "markers.toml")),
+    "PATH_KB_CONFIG": str(Path(Path.home(), ".kb", "kb.conf.py")),  # for future use
+    "PATH_KB_TEMPLATES": str(Path(Path.home(), ".kb", "templates")),
+    "PATH_KB_DEFAULT_TEMPLATE": str(Path(Path.home(), ".kb", "templates", "default")),
+    "DB_SCHEMA_VERSION": 1,
     "EDITOR": os.environ.get("EDITOR", "vim"),
-    "INITIAL_CATEGORIES": ["default",
-                           "cheatsheets",
-                           "procedures",
-                           "guides"],
+    "INITIAL_CATEGORIES": ["default",]
 }
 
 
-DEFAULT_MARKERS = {
+DEFAULT_TEMPLATE = {
     "TITLES": ("^#.*", "blue"),
     "WARNINGS": ("^!.*", "yellow"),
 }