+++ CHANGELOG.md
@@ -25,3 +25,8 @@
     * Improved continuous integration
 Draft 0.1.4
     * Added body "-b" function in add
+    * Fixed bugs related to the update function
+    * Added templates
+    * Added schema migration logic
+    * Updated docs with a TOC
+