+++ docs/CHANGELOG.md
@@ -3,7 +3,7 @@
 :Author: gnc <nebbionegiuseppe@gmail.com>
 :Copyright: © 2020, gnc.
 :License: GPLv3 (see /LICENSE or :doc:`Appendix B <LICENSE>`.)
-:Date: 2020-08-09
+:Date: 2020-09-29
 :Version: 0.1.1
 
 ## Version History
@@ -23,4 +23,10 @@
     * Fixed Windows bugs
     * Fixed cross-platforms tests
     * Improved continuous integration
+Draft 0.1.4
+    * Added body "-b" function in add
+    * Fixed bugs related to the update function
+    * Added templates
+    * Added schema migration logic
+    * Updated docs with a TOC